+++ docker/env_file_app_ci
@@ -124,5 +124,3 @@ COVERAGE_PROCESS_START=.coveragerc
 # broker configuration
 BROKER_URL=amqp://guest:guest@rabbitmq:5672
 
-# Playbooks
-TEST_PLAYBOOK=FREE_TO_USE_ANALYZERS
