+++ docker/env_file_app_template
@@ -73,5 +73,3 @@ RABBITMQ_DEFAULT_PASS=guest
 FLOWER_USER=flower
 FLOWER_PWD=flower
 
-# playbooks
-TEST_PLAYBOOK=FREE_TO_USE_ANALYZERS
